Moved Debug to helpers
