+++ src/helpers/favicon.ts
@@ -1,5 +1,5 @@
 import { HtmlUtils } from './HtmlUtils';
-import { Debug } from '../Debug';
+import { Debug } from './Debug';
 import { safeFetch } from './safeFetch';
 import * as urlUtils from './urlUtils';
