+++ src/helpers/opmlImport.ts
@@ -1,7 +1,7 @@
 import { Feed } from '../models/Feed';
 import { RSSFeedManager } from '../RSSPostManager';
 import { getHttpsUrl } from './urlUtils';
-import { Debug } from '../Debug';
+import { Debug } from './Debug';
 import { timeout } from './Utils';
 
 // tslint:disable-next-line: no-var-requires
