+++ src/ui/screens/bug-report/BugReportScreen.tsx
@@ -17,7 +17,7 @@ import { BoldText } from '../../misc/text';
 import { filteredLog } from '../../../helpers/log';
 import { Version, BuildNumber } from '../../../Version';
 
-import { Debug } from '../../../Debug';
+import { Debug } from '../../../helpers/Debug';
 import { TypedNavigation } from '../../../helpers/navigation';
 import { SimpleTextInput } from '../../misc/SimpleTextInput';
 import { WideButton } from '../../buttons/WideButton';
